ai/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0

# ML
torch==2.1.0
transformers==4.35.0
pillow==10.1.0
opencv-python-headless==4.8.1.78
numpy==1.24.3
scikit-image==0.22.0

# LLM & utilities
openai==1.3.0
langchain==0.0.340
langchain-openai==0.0.2
tiktoken==0.5.1
anthropic==0.7.0

# AWS
boto3==1.29.0
botocore==1.32.0
aioboto3==12.0.0

# Graph DB
gremlinpython==3.7.0
nest-asyncio==1.5.8

# DB
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1

# Caching
redis==5.0.1
hiredis==2.2.3

# Utils
python-dotenv==1.0.0
python-json-logger==2.0.7
requests==2.31.0
aiohttp==3.9.0
tenacity==8.2.3
pydantic-settings==2.1.0

# Dev (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.0
